Multilayer Network with Backpropagation
The goal of this project is to train a multilayer perceptron that uses backpropagation. Once 3 sets are trained, the best one amongst them is picked to test the hit ratio. Afterwards the fault tolerance of the best set is tested by setting 20% of the weights to zero, and then tested again by setting another 20% of the weights to zero.